refactor(1013): remove SiteAnomalyExporter facade (SC-001, position 43)
Extract SiteAnomalyExporter (54 refs, 341 LOC) to src/export/site_anomaly_exporter.py
under the SC-001 facade pattern. Lazy mh = importlib.import_module("MistHelper") resolves
every live global (PromptUtils, EnhancedSSHRunner, AnomalyMetricsDiscovery, mistapi,
apisession, DataProcessingUtils, DataExporter, PromptClientUtils) inside each helper.

- Add re-export alias in MistHelper.py alphabetically among src.export.* imports
- Replace class body with single-line breadcrumb (avoids stub-class shadow)
- Preserve AnomalyMetricsDiscovery import with F401 breadcrumb (lazy access via mh.*)
- Add coverage-omit entry for src/export/site_anomaly_exporter.py
- src.export.* already in mypy Wave 2 override; no mypy change needed

Local gate: black + ruff clean; python MistHelper.py --test = 66/0/131 baseline.
