+++ MistHelper.py
@@ -178,6 +178,9 @@ from src.export.org_template_exporter import (  # pylint: disable=unused-import
 from src.export.self_export_utils import (  # pylint: disable=unused-import
     SelfExportUtils,  # noqa: F401  # Cat B (1013 SC-001 position 7) -- re-export for menu tuple at MistHelper:18167
 )
+from src.export.site_anomaly_exporter import (  # pylint: disable=unused-import
+    SiteAnomalyExporter,  # noqa: F401  # Cat B (1013 SC-001 position 43) -- re-export for MistHelper.SiteAnomalyExporter callers
+)
 from src.export.site_client_exporter import (  # pylint: disable=unused-import
     SiteClientExporter,  # noqa: F401  # Cat B (1013 SC-001 position 14) -- re-export for MistHelper.SiteClientExporter callers
 )
@@ -218,8 +221,8 @@ from src.inventory.org_device_inventory_summary_facade import (  # pylint: disab
 )
 from src.org.org_config_migration_manager import OrgConfigMigrationManager  # Cat B (1013 SC-001 position 5)
 from src.org_data_collector import OrgDataCollector  # Import org-level data collection orchestrator
-from src.refactors.anomaly_metrics_discovery import (
-    AnomalyMetricsDiscovery,  # Extracted anomaly metrics discovery (SC-016)
+from src.refactors.anomaly_metrics_discovery import (  # pylint: disable=unused-import
+    AnomalyMetricsDiscovery,  # noqa: F401  # Cat B (1013 SC-001 pos 43) -- lazy access via mh.AnomalyMetricsDiscovery
 )
 from src.refactors.connection_pool_executor import ConnectionPoolExecutor  # Extracted pool executor (1012 SC-003)
 from src.refactors.data_directory_checker import DataDirectoryChecker  # Early data-dir writable check (SC-005)
@@ -9764,347 +9767,8 @@ class OrgExportUtils:  # Generic org export helpers.
 # SiteConfigExporter moved to src/export/site_config_exporter.py (1013 SC-001 position 19)
 
 
-class SiteAnomalyExporter:  # Site anomaly exporters.
-    """
-    Site Anomaly and Event Exporter
-
-    Handles site-level anomaly events and insight metrics exports.
-    Extracted from SiteExportUtils.
-    """
-
-    @staticmethod
-    def anomaly_events():
-        """Export comprehensive anomaly events for a selected site to SiteAnomalyEvents_[SiteName].csv."""
-        print("Export Site Anomaly Events:")  # User-visible header.
-        logging.info("Starting export of site anomaly events...")  # Trace start of export.
-        site_id = PromptUtils.select_site()  # Prompt the user for a site.
-        if not site_id:  # No site chosen.
-            print("! No site selected. Exiting.")  # Tell the user.
-            return  # Abort the export.
-        site_name = SiteAnomalyExporter._anomaly_resolve_site_name(site_id)  # Resolve display name for filename.
-        filename = f"SiteAnomalyEvents_{EnhancedSSHRunner.sanitize_filename(site_name)}.csv"  # Build CSV name.
-        metrics = SiteAnomalyExporter._discover_site_anomaly_metrics()  # Discover anomaly metric names.
-        if not metrics:  # Nothing to fetch.
-            return  # Abort the export.
-        try:
-            data, count = SiteAnomalyExporter._aggregate_site_anomaly_data(site_id, site_name, metrics)  # Fetch.
-            SiteAnomalyExporter._export_anomaly_data(data, filename, "site anomaly event", count, site_name)  # CSV
-        except Exception as exception:  # Broader export failure (flatten/write).
-            print(f"! Error exporting site anomaly events: {exception}")  # Tell the user.
-            logging.error("Failed to export site anomaly events for %s: %s", site_name, exception)  # Log it.
-
-    @staticmethod
-    def device_anomaly_events():
-        """Export device anomaly events to SiteDeviceAnomalyEvents_[Site]_[Device].csv."""
-        print("Export Site Device Anomaly Events:")  # User-visible header.
-        logging.info("Starting export of site device anomaly events...")  # Trace start of export.
-        site_id = PromptUtils.select_site()  # Prompt the user for a site.
-        if not site_id:  # No site chosen.
-            print("! No site selected. Exiting.")  # Tell the user.
-            return  # Abort the export.
-        site_name = SiteAnomalyExporter._anomaly_resolve_site_name(site_id)  # Resolve display name.
-        selection = PromptUtils.select_device_id_from_inventory(site_id)  # Prompt for a device.
-        if not selection:  # No device chosen.
-            print("! No device selected. Exiting.")  # Tell the user.
-            return  # Abort the export.
-        device_mac, device_name = selection[0], selection[1]  # Unpack the MAC and display name.
-        filename = SiteAnomalyExporter._build_device_filename(site_name, device_name)  # Build the CSV name.
-        metrics = ["ap_availability", "throughput", "capacity"]  # Device anomaly metric names.
-        try:
-            data, count = SiteAnomalyExporter._aggregate_device_anomaly_data(
-                site_id, site_name, device_mac, device_name, metrics
-            )  # Loop + fetch each metric.
-            SiteAnomalyExporter._export_anomaly_data(data, filename, "device anomaly event", count, device_name)  # CSV
-        except Exception as exception:  # Broader export failure (flatten/write).
-            print(f"! Error exporting device anomaly events: {exception}")  # Tell the user.
-            logging.error("Failed to export device anomaly events for %s: %s", device_name, exception)  # Log it.
-
-    @staticmethod
-    def _build_device_filename(site_name: str, device_name: str) -> str:
-        """Build the device anomaly CSV filename from sanitized site + device names."""
-        sanitized_site = EnhancedSSHRunner.sanitize_filename(site_name)  # Sanitize the site name.
-        sanitized_device = EnhancedSSHRunner.sanitize_filename(device_name)  # Sanitize the device name.
-        return f"SiteDeviceAnomalyEvents_{sanitized_site}_{sanitized_device}.csv"  # Compose CSV name.
-
-    @staticmethod
-    def _discover_site_anomaly_metrics() -> list[str]:
-        """Discover potential site anomaly metric names, announce them, return [] when none found."""
-        print("! Discovering potential anomaly metrics from Mist API definitions...")  # Tell the user.
-        potential = AnomalyMetricsDiscovery.discover()  # Pull discovery list from CSV.
-        names = [info["metric_name"] for info in potential]  # Extract just the metric names.
-        print(f"! Found {len(names)} potential anomaly metrics:")  # Tell the user the count.
-        for info in potential:  # Show each metric to the user.
-            print(f"  - {info['metric_name']}: {info['description'][:60]}...")  # Trim long descriptions.
-        if not names:  # No metrics discovered at all.
-            print("! No potential anomaly metrics found. Please check ConstInsightMetrics.csv availability.")
-        return names  # Return the names (possibly empty).
-
-    @staticmethod
-    def _fetch_one_anomaly_metric(
-        fetch_callable: Any, metric: str, tags: dict[str, str], scope: tuple[str, str]
-    ) -> dict[str, Any] | None:
-        """Invoke fetch_callable() for one metric, tag the result, return data dict or None on no-data/error."""
-        display_label, data_type = scope  # Unpack the scope tuple for printing + tagging.
-        try:
-            response = fetch_callable()  # Issue the API call (callable is already bound via functools.partial).
-            data = getattr(response, "data", response) or {}  # Unwrap data; default empty.
-            if data:  # API returned actual data for this metric.
-                data["metric_type"] = metric  # Tag the metric name.
-                data["data_type"] = data_type  # Tag the data type for downstream consumers.
-                for key, value in tags.items():  # Apply caller-supplied tags (site_id, site_name, ...).
-                    data[key] = value  # Set each tag on the result row.
-                print(f"!? Retrieved {metric} {display_label}")  # Tell the user.
-                logging.debug("Successfully retrieved %s %s for %s", metric, display_label, tags)  # Trace success.
-                return data  # Return the tagged row.
-            print(f"! No {metric} {display_label} available")  # Tell the user the metric had no data.
-            logging.info("No %s %s available for %s", metric, display_label, tags)  # Log the absence.
-            return None  # Signal caller to skip.
-        except Exception as metric_error:  # This metric's fetch failed.
-            print(f"! Error retrieving {metric} {display_label}: {metric_error}")  # Tell the user.
-            logging.warning("Error retrieving %s %s for %s: %s", metric, display_label, tags, metric_error)  # Warn.
-            return None  # Signal caller to skip.
-
-    @staticmethod
-    def _run_anomaly_metric_loop(
-        metrics: list[str],
-        fetch_builder: Callable[[str], Callable],
-        tags: dict[str, Any],
-        scope: tuple[str, str],
-    ) -> tuple[list[dict[str, Any]], int]:
-        """Loop metrics with mistapi loggers silenced; build fetch with fetch_builder(metric); collect tagged rows."""
-        original_levels = SiteAnomalyExporter._anomaly_suppress_mistapi_loggers()  # Quiet mistapi internal loggers.
-        rows: list[dict[str, Any]] = []  # Accumulator for successful rows.
-        count = 0  # Number of metrics that returned data.
-        try:
-            for metric in metrics:  # Fetch each metric.
-                fetch = fetch_builder(metric)  # Build per-metric fetch callable.
-                row = SiteAnomalyExporter._fetch_one_anomaly_metric(fetch, metric, tags, scope)  # Fetch + tag.
-                if row is not None:  # Metric returned data.
-                    rows.append(row)  # Collect the row.
-                    count += 1  # Bump the success counter.
-        finally:
-            SiteAnomalyExporter._anomaly_restore_loggers(original_levels)  # Always restore loggers.
-        return rows, count  # Hand the aggregate back to the orchestrator.
-
-    @staticmethod
-    def _aggregate_site_anomaly_data(
-        site_id: str, site_name: str, metrics: list[str]
-    ) -> tuple[list[dict[str, Any]], int]:
-        """Loop site anomaly metrics with mistapi loggers silenced; return (rows, success_count)."""
-        tags = {"site_id": site_id, "site_name": site_name}  # Tags attached to every row.
-        scope = ("anomaly events", "site_anomaly_events")  # Display label + data_type tag.
-        print(f"! Retrieving {len(metrics)} different site anomaly events...")  # Tell the user.
-        builder = lambda metric: functools.partial(  # noqa: E731 — bind site-anomaly fetch per metric.
-            mistapi.api.v1.sites.anomaly.listSiteAnomalyEvents, apisession, site_id, metric
-        )
-        return SiteAnomalyExporter._run_anomaly_metric_loop(metrics, builder, tags, scope)  # Shared loop.
-
-    @staticmethod
-    def _aggregate_device_anomaly_data(
-        site_id: str, site_name: str, device_mac: str, device_name: str, metrics: list[str]
-    ) -> tuple[list[dict[str, Any]], int]:
-        """Loop device anomaly metrics with mistapi loggers silenced; return (rows, success_count)."""
-        tags = {  # Tags attached to every row.
-            "site_id": site_id,
-            "site_name": site_name,
-            "device_mac": device_mac,
-            "device_name": device_name,
-        }
-        scope = ("device anomaly data", "device_anomaly_events")  # Display label + data_type tag.
-        print(f"! Retrieving {len(metrics)} different device anomaly events for {device_name}...")  # Tell the user.
-        builder = lambda metric: functools.partial(  # noqa: E731 — bind device-anomaly fetch per metric.
-            mistapi.api.v1.sites.anomaly.getSiteAnomalyEventsForDevice, apisession, site_id, metric, device_mac
-        )
-        return SiteAnomalyExporter._run_anomaly_metric_loop(metrics, builder, tags, scope)  # Shared loop.
-
-    @staticmethod
-    def _export_anomaly_data(
-        data_list: list[dict[str, Any]], filename: str, label: str, success_count: int, scope_name: str
-    ) -> None:
-        """Flatten + escape + write the aggregated anomaly rows, or write an empty CSV when there is no data."""
-        if data_list:  # At least one metric returned data.
-            processed = DataProcessingUtils.flatten_nested_fields(data_list)  # Flatten nested fields.
-            processed = DataProcessingUtils.escape_multiline(processed)  # type: ignore[no-untyped-call]
-            DataExporter.write_with_format_selection(processed, filename)  # type: ignore[no-untyped-call]
-            print(f"! {success_count} {label} types exported to {filename}")  # Tell the user the count.
-            logging.info("Exported %s %s types for %s to %s", success_count, label, scope_name, filename)  # Log.
-        else:  # No data from any metric.
-            print(f"! 0 {label}s exported to {filename} (no data available)")  # Tell the user zero.
-            logging.warning("No %s available for %s", label, scope_name)  # Warn about the empty result.
-            DataExporter.write_with_format_selection([], filename)  # type: ignore[no-untyped-call]
-
-    _CLIENT_ANOMALY_METRICS = (  # Client-specific anomaly metrics (verified working) shared by the count + loop.
-        "successful_connect",  # Note: uses underscore, not hyphen for the client endpoint.
-        "roaming",  # Client roaming issues.
-        "throughput",  # Client throughput anomalies.
-    )
-
-    @staticmethod
-    def _anomaly_resolve_site_name(site_id: str) -> str:
-        """Resolve the human-readable site name for a site_id, falling back to the id on lookup failure."""
-        try:  # The site-name lookup is best-effort; the id is an acceptable fallback for the filename.
-            response = mistapi.api.v1.sites.listSites(apisession, site_id)  # List the site.
-            sites = mistapi.get_all(response=response, mist_session=apisession)  # Page all rows.
-            return next((site["name"] for site in sites if site["id"] == site_id), site_id)  # Resolve site name.
-        except Exception:  # Lookup failed.
-            return site_id  # Fall back to the id.
-
-    @staticmethod
-    def _anomaly_lookup_client_hostname(site_id: str, client_mac: str) -> str:
-        """Look up a client's hostname from its wireless stats, falling back to the MAC on failure."""
-        try:  # Hostname enrichment is best-effort; the MAC is an acceptable fallback.
-            response = mistapi.api.v1.sites.stats.listSiteWirelessClientsStats(  # List client stats.
-                apisession, site_id, limit=100, duration="1d"
-            )
-            clients = getattr(response, "data", response) or []  # Unwrap data; default empty.
-            for client in clients:  # Find the matching client.
-                if client.get("mac") == client_mac:  # MAC matches.
-                    return client.get("hostname", client.get("name", "Unknown"))  # Read the hostname.
-            return "Unknown"  # No matching client found in the stats.
-        except Exception as exception:  # Lookup failed.
-            logging.warning("Could not retrieve client hostname for %s: %s", client_mac, exception)  # Warn the failure.
-            return client_mac  # Fall back to the MAC address.
-
-    @staticmethod
-    def _anomaly_suppress_mistapi_loggers() -> dict:  # type: ignore[type-arg]
-        """Raise mistapi logger levels to CRITICAL to keep the console clean; return their original levels."""
-        mistapi_loggers = [
-            "apirequest",
-            "apiresponse",
-            "mistapi",
-            "mistapi.apirequest",
-            "mistapi.apiresponse",
-        ]  # Names.
-        original_levels = {}  # Save original levels for later restoration.
-        for logger_name in mistapi_loggers:  # Quiet each mistapi logger.
-            logger_instance = logging.getLogger(logger_name)  # Get the logger.
-            original_levels[logger_name] = logger_instance.level  # Remember its level.
-            logger_instance.setLevel(logging.CRITICAL)  # Suppress ERROR logs temporarily.
-        return original_levels  # Original levels keyed by logger name.
-
-    @staticmethod
-    def _anomaly_restore_loggers(original_levels: dict) -> None:  # type: ignore[type-arg]
-        """Restore mistapi logger levels saved by _anomaly_suppress_mistapi_loggers."""
-        for logger_name, original_level in original_levels.items():  # Restore each logger level.
-            logging.getLogger(logger_name).setLevel(original_level)  # Restore the saved level.
-
-    @staticmethod
-    def _anomaly_fetch_one_metric(
-        site_id: str, client_mac: str, site_name: str, client_hostname: str, metric: str
-    ) -> dict | None:  # type: ignore[type-arg]
-        """Fetch one client anomaly metric and tag it with site/client metadata; return the record, or None if empty."""
-        response = mistapi.api.v1.sites.anomaly.getSiteAnomalyEventsForClient(  # Get client anomalies for this metric.
-            apisession, site_id, client_mac, metric
-        )
-        client_anomaly_data = getattr(response, "data", response) or {}  # Unwrap data; default empty.
-        if not client_anomaly_data:  # The metric returned no data.
-            return None  # Signal an empty metric.
-        client_anomaly_data["metric_type"] = metric  # Tag the metric.
-        client_anomaly_data["site_id"] = site_id  # Tag the site.
-        client_anomaly_data["site_name"] = site_name  # Tag the site name.
-        client_anomaly_data["client_mac"] = client_mac  # Tag the client MAC.
-        client_anomaly_data["client_hostname"] = client_hostname  # Tag the hostname.
-        client_anomaly_data["data_type"] = "client_anomaly_events"  # Tag the data type.
-        return client_anomaly_data  # The tagged anomaly record.
-
-    @staticmethod
-    def _anomaly_handle_metric_result(record: dict | None, metric: str, client_mac: str, all_data: list) -> int:  # type: ignore[type-arg]
-        """Record one fetched metric: append + announce on data, announce 'none' otherwise; return 1 if kept else 0."""
-        if record is not None:  # The metric returned data.
-            all_data.append(record)  # Collect the row.
-            print(f"!? Retrieved {metric} client anomaly data")  # Tell the user.
-            logging.debug("Successfully retrieved %s client anomaly data for %s", metric, client_mac)  # Trace.
-            return 1  # One successful metric.
-        print(f"! No {metric} client anomaly data available")  # Tell the user none.
-        logging.info("No %s client anomaly data available for %s", metric, client_mac)  # Log none.
-        return 0  # No data for this metric.
-
-    @staticmethod
-    def _anomaly_collect_metrics(
-        site_id: str, client_mac: str, site_name: str, client_hostname: str
-    ) -> tuple[list, int]:  # type: ignore[type-arg]
-        """Fetch all client anomaly metrics for one client; return (records, retrieved_count)."""
-        print(
-            f"! Retrieving {len(SiteAnomalyExporter._CLIENT_ANOMALY_METRICS)} different client anomaly events for {client_mac} ({client_hostname})..."  # noqa: E501
-        )  # Tell the user how many metrics will be fetched.
-        all_client_anomaly_data = []  # Accumulate anomaly rows.
-        metrics_retrieved = 0  # Success count.
-        for metric in SiteAnomalyExporter._CLIENT_ANOMALY_METRICS:  # Fetch each metric independently.
-            try:  # Isolate per-metric failures so one bad metric doesn't abort the rest.
-                record = SiteAnomalyExporter._anomaly_fetch_one_metric(  # Fetch + tag one metric.
-                    site_id, client_mac, site_name, client_hostname, metric
-                )
-                metrics_retrieved += SiteAnomalyExporter._anomaly_handle_metric_result(  # Record + announce outcome.
-                    record, metric, client_mac, all_client_anomaly_data
-                )
-            except Exception as metric_error:  # Metric fetch failed.
-                print(f"! Error retrieving {metric} client anomaly data: {metric_error}")  # Tell the user.
-                logging.warning(
-                    "Error retrieving %s client anomaly data for %s: %s", metric, client_mac, metric_error
-                )  # Warn the failure.
-        return all_client_anomaly_data, metrics_retrieved  # Collected rows and the success count.
-
-    @staticmethod
-    def _anomaly_export(all_data: list, metrics_retrieved: int, client_mac: str, filename: str) -> None:  # type: ignore[type-arg]
-        """Flatten and write the collected client anomaly rows to CSV (writes an empty file when there is no data)."""
-        if all_data:  # At least one metric returned data.
-            processed = DataProcessingUtils.flatten_nested_fields(all_data)  # Flatten nested fields.
-            processed = DataProcessingUtils.escape_multiline(processed)  # type: ignore[no-untyped-call]
-            DataExporter.write_with_format_selection(processed, filename)  # type: ignore[no-untyped-call]
-            print(f"! {metrics_retrieved} client anomaly event types exported to {filename}")  # Tell the user.
-            logging.info(
-                "Exported %s client anomaly event types for %s to %s", metrics_retrieved, client_mac, filename
-            )  # Log the export.
-        else:  # No metric returned data.
-            print(f"! 0 client anomaly events exported to {filename} (no data available)")  # Tell the user zero.
-            logging.warning("No client anomaly events available for %s", client_mac)  # Warn none.
-            DataExporter.write_with_format_selection([], filename)  # type: ignore[no-untyped-call]  # Write empty file.
-
-    @staticmethod
-    def _anomaly_prepare() -> tuple | None:  # type: ignore[type-arg]
-        """Prompt for a site and client, resolve names + hostname, and build the output filename.
-
-        Returns (site_id, site_name, client_mac, client_hostname, filename), or None when the
-        operator cancels at the site or client selection prompt.
-        """
-        site_id = PromptUtils.select_site()  # Select a site.
-        if not site_id:  # No site selected.
-            print("! No site selected. Exiting.")  # Tell the user.
-            return None  # Abort.
-        site_name = SiteAnomalyExporter._anomaly_resolve_site_name(site_id)  # Resolve site name for the filename.
-        client_mac, _, _ = PromptClientUtils.select_client(site_id)  # Select a client (only MAC is used).
-        if not client_mac:  # No client selected.
-            print("! No client selected. Exiting.")  # Tell the user.
-            return None  # Abort.
-        client_hostname = SiteAnomalyExporter._anomaly_lookup_client_hostname(site_id, client_mac)  # Hostname lookup.
-        sanitized_site_name = EnhancedSSHRunner.sanitize_filename(site_name)  # Sanitize the site name.
-        filename = f"SiteClientAnomalyEvents_{sanitized_site_name}_{client_mac.replace(':', '')}.csv"  # Output name.
-        return site_id, site_name, client_mac, client_hostname, filename  # Resolved context for the export.
-
-    @staticmethod
-    def client_anomaly_events():
-        """Export client-specific anomaly events for a selected client to a per-site/per-client CSV.
-
-        Prompts for a site and client, fetches the successful_connect / roaming / throughput
-        anomaly metrics, and writes SiteClientAnomalyEvents_[SiteName]_[ClientMAC].csv.
-        """
-        print("Export Site Client Anomaly Events:")  # Header.
-        logging.info("Starting export of site client anomaly events...")  # Log start.
-        prepared = SiteAnomalyExporter._anomaly_prepare()  # Prompt + resolve site/client/filename.
-        if prepared is None:  # Operator cancelled at a selection prompt.
-            return  # Abort.
-        site_id, site_name, client_mac, client_hostname, filename = prepared  # Unpack the resolved context.
-        original_levels = SiteAnomalyExporter._anomaly_suppress_mistapi_loggers()  # Quiet mistapi loggers.
-        try:  # Guard the fetch/export so logger levels are always restored in finally.
-            all_data, metrics_retrieved = SiteAnomalyExporter._anomaly_collect_metrics(  # Fetch all metrics.
-                site_id, client_mac, site_name, client_hostname
-            )
-            SiteAnomalyExporter._anomaly_export(all_data, metrics_retrieved, client_mac, filename)  # Write the CSV.
-        except Exception as exception:  # Export failed.
-            print(f"! Error exporting client anomaly events: {exception}")  # Tell the user.
-            logging.error("Failed to export client anomaly events for %s: %s", client_mac, exception)  # Log the error.
-        finally:  # Always restore the mistapi logger levels.
-            SiteAnomalyExporter._anomaly_restore_loggers(original_levels)  # Restore logger levels.
+# --- SiteAnomalyExporter facade removed (1013 SC-001 Cat B pos 43) ---
+# Canonical implementation lives in src/export/site_anomaly_exporter.py; re-exported above.
 
 
 # --- SitesByAPModelExporter facade removed (1013 SC-001 Cat B pos 28) ---
